+++ src/components/NoticeList/NoticeList.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-	import useGetNoticeListQuery from '@/hooks/useGetNoticeList.hook';
+	import useGetNoticeListQuery from '@/hooks/notice/useGetNoticeList.hook';
 	import NoticeItem from './NoticeItem.svelte';
 
 	const query = useGetNoticeListQuery();
